pyproject.toml
[tool.poetry]
name        = "blog"
version     = "1.0.0"
description = "Hyperreals *R"
authors     = ["chemacortes <devxtrem@gmail.com>"]

[tool.poetry.dependencies]
python      = "^3.9"
pelican     = "^4.5"
webassets   = "^2.0"
markdown    = "^3.3"
typogrify   = "^2.0.7"
cssmin      = "^0.2.0"
ghp-import2 = "^1.0.1"
slimit      = "^0.8.1"
pyscss      = "^1.3.7"
Jinja2      = "^2.11.2"
mdx-include = "^1.4.1"
ply         = "^3.11"

[tool.poetry.dev-dependencies]
pylama         = "^7.7.1"
black          = { version = "^20.8b1", allow-prereleases = true }
python-slugify = "^4.0.1"
invoke         = "^1.5.0"
ipykernel = "^5.4.3"
coconut = "^1.4.3"

[build-system]
requires      = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.poetry.scripts]
slimit_warning = "blog.scripts:remove_slimit_warning"
coconut_lexer = "blog.scripts:install_lexer_coconut"